Int is now 64 bits

+++ src/core/java/panda/core/Int32Wrapper_ExternalImpl.java
@@ -2,7 +2,7 @@ package panda.core;
 
 class Int32Wrapper_ExternalImpl implements panda.core.Int32Wrapper_External {
     @Override
-    public int get_length_$Rpanda$core$Int32(int self) {
-        return 32 - java.lang.Integer.numberOfLeadingZeros(self);
+    public byte get_length_$Rpanda$core$UInt8(int self) {
+        return (byte) (32 - java.lang.Integer.numberOfLeadingZeros(self));
     }
 }
